reviewer's feedback applied
