+++ modules/plugin/imagemosaic/src/main/java/org/geotools/gce/imagemosaic/catalog/FeatureTypeMapper.java
@@ -14,7 +14,7 @@
  *    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
  *    Lesser General Public License for more details.
  */
-package org.geotools.gce.imagemosaic.catalog.oracle;
+package org.geotools.gce.imagemosaic.catalog;
 
 import java.util.List;
 import org.geotools.data.simple.SimpleFeatureSource;
