+++ modules/plugin/imagemosaic/src/main/java/org/geotools/gce/imagemosaic/catalog/oracle/OracleDatastoreWrapper.java
@@ -20,6 +20,8 @@ import java.io.IOException;
 import org.geotools.data.DataStore;
 import org.geotools.data.simple.SimpleFeatureSource;
 import org.geotools.data.simple.SimpleFeatureStore;
+import org.geotools.gce.imagemosaic.catalog.DataStoreWrapper;
+import org.geotools.gce.imagemosaic.catalog.FeatureTypeMapper;
 import org.opengis.feature.simple.SimpleFeatureType;
 
 /**
