+++ modules/plugin/imagemosaic/src/main/java/org/geotools/gce/imagemosaic/catalog/postgis/PostgisDatastoreWrapper.java
@@ -23,8 +23,8 @@ import org.geotools.data.DataStore;
 import org.geotools.data.simple.SimpleFeatureSource;
 import org.geotools.data.simple.SimpleFeatureStore;
 import org.geotools.data.transform.Definition;
-import org.geotools.gce.imagemosaic.catalog.oracle.DataStoreWrapper;
-import org.geotools.gce.imagemosaic.catalog.oracle.FeatureTypeMapper;
+import org.geotools.gce.imagemosaic.catalog.DataStoreWrapper;
+import org.geotools.gce.imagemosaic.catalog.FeatureTypeMapper;
 import org.opengis.feature.simple.SimpleFeatureType;
 
 /**
